+++ lightly_studio/tests/dataset/test_embedding_manager.py
@@ -2,6 +2,7 @@
 
 from __future__ import annotations
 
+from typing import Any
 from uuid import UUID, uuid4
 
 import numpy as np
@@ -103,7 +104,10 @@ def test_register_multiple_models(
             raise NotImplementedError()
 
         def embed_image_crops(
-            self, image_crops: list[ImageCrop], show_progress: bool = True
+            self,
+            image_crops: list[ImageCrop],
+            show_progress: bool = True,
+            progress_bar: Any | None = None,
         ) -> NDArray[np.float32]:
             raise NotImplementedError()
